Refactors product page; adds deleted item view
Extracts interactive product listing and form logic into a dedicated client component (`page-client.tsx`). This aligns with Next.js App Router best practices, separating client-side interactivity from the server page.

Introduces a new feature to view soft-deleted products via a "Show deleted items" checkbox, enhancing data visibility. Also updates the page title and adds metadata.
